+++ orm/model/ormtable/bench_test.go
@@ -11,6 +11,7 @@ import (
 	"gotest.tools/v3/assert"
 
 	"cosmossdk.io/core/store"
+	coretesting "cosmossdk.io/core/testing"
 	"cosmossdk.io/orm/internal/testkv"
 	"cosmossdk.io/orm/internal/testpb"
 	"cosmossdk.io/orm/model/ormtable"
@@ -241,7 +242,7 @@ func getBalance(store kv.Store, address, denom string) (*testpb.Balance, error) 
 
 func BenchmarkManualInsertMemory(b *testing.B) {
 	benchManual(b, func() (store.KVStore, error) {
-		return testkv.TestStore{Db: dbm.NewMemDB()}, nil
+		return testkv.TestStore{Db: coretesting.NewMemDB()}, nil
 	})
 }