+++ runtime/v2/builder.go
@@ -6,28 +6,22 @@ import (
 	"errors"
 	"fmt"
 	"io"
-	"path/filepath"
 
 	"cosmossdk.io/core/appmodule"
 	appmodulev2 "cosmossdk.io/core/appmodule/v2"
-	"cosmossdk.io/core/server"
 	"cosmossdk.io/core/store"
 	"cosmossdk.io/core/transaction"
 	"cosmossdk.io/runtime/v2/services"
 	"cosmossdk.io/server/v2/appmanager"
 	"cosmossdk.io/server/v2/stf"
 	"cosmossdk.io/server/v2/stf/branch"
-	"cosmossdk.io/store/v2/db"
-	rootstore "cosmossdk.io/store/v2/root"
 )
 
 // AppBuilder is a type that is injected into a container by the runtime/v2 module
 // (as *AppBuilder) which can be used to create an app which is compatible with
 // the existing app.go initialization conventions.
 type AppBuilder[T transaction.Tx] struct {
-	app          *App[T]
-	config       server.DynamicConfig
-	storeOptions *rootstore.Options
+	app *App[T]
 
 	// the following fields are used to overwrite the default
 	branch      func(state store.ReaderMap) store.WriterMap
@@ -99,6 +93,10 @@ func (a *AppBuilder[T]) Build(opts ...AppBuilderOption[T]) (*App[T], error) {
 		}
 	}
 
+	if a.app.db == nil {
+		return nil, fmt.Errorf("app.db is not set, it is required to build the app")
+	}
+
 	if err := a.app.moduleManager.RegisterServices(a.app); err != nil {
 		return nil, err
 	}
@@ -121,37 +119,6 @@ func (a *AppBuilder[T]) Build(opts ...AppBuilderOption[T]) (*App[T], error) {
 		return nil, fmt.Errorf("failed to create STF: %w", err)
 	}
 	a.app.stf = stf
-
-	home := a.config.GetString(FlagHome)
-	scRawDb, err := db.NewDB(
-		db.DBType(a.config.GetString("store.app-db-backend")),
-		"application",
-		filepath.Join(home, "data"),
-		nil,
-	)
-	if err != nil {
-		panic(err)
-	}
-
-	var storeOptions rootstore.Options
-	if a.storeOptions != nil {
-		storeOptions = *a.storeOptions
-	} else {
-		storeOptions = rootstore.DefaultStoreOptions()
-	}
-	factoryOptions := &rootstore.FactoryOptions{
-		Logger:    a.app.logger,
-		RootDir:   home,
-		Options:   storeOptions,
-		StoreKeys: append(a.app.storeKeys, "stf"),
-		SCRawDB:   scRawDb,
-	}
-
-	rs, err := rootstore.CreateRootStore(factoryOptions)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create root store: %w", err)
-	}
-	a.app.db = rs
 
 	appManagerBuilder := appmanager.Builder[T]{
 		STF:                a.app.stf,
@@ -193,17 +160,17 @@ func (a *AppBuilder[T]) Build(opts ...AppBuilderOption[T]) (*App[T], error) {
 			return genesisState, err
 		},
 		ExportGenesis: func(ctx context.Context, version uint64) ([]byte, error) {
-			_, state, err := a.app.db.StateLatest()
+			state, err := a.app.db.StateAt(version)
 			if err != nil {
-				return nil, fmt.Errorf("unable to get latest state: %w", err)
+				return nil, fmt.Errorf("unable to get state at given version: %w", err)
 			}
-			genesisCtx := services.NewGenesisContext(a.branch(state))
 
-			var genesisJson map[string]json.RawMessage
-			_, err = genesisCtx.Run(ctx, func(ctx context.Context) error {
-				genesisJson, err = a.app.moduleManager.ExportGenesisForModules(ctx)
-				return err
-			})
+			genesisJson, err := a.app.moduleManager.ExportGenesisForModules(
+				ctx,
+				func() store.WriterMap {
+					return a.branch(state)
+				},
+			)
 			if err != nil {
 				return nil, fmt.Errorf("failed to export genesis: %w", err)
 			}
@@ -249,11 +216,5 @@ func AppBuilderWithTxValidator[T transaction.Tx](txValidators func(ctx context.C
 func AppBuilderWithPostTxExec[T transaction.Tx](postTxExec func(ctx context.Context, tx T, success bool) error) AppBuilderOption[T] {
 	return func(a *AppBuilder[T]) {
 		a.postTxExec = postTxExec
-	}
-}
-
-func AppBuilderWithStoreOptions[T transaction.Tx](opts *rootstore.Options) AppBuilderOption[T] {
-	return func(a *AppBuilder[T]) {
-		a.storeOptions = opts
 	}
 }