+++ x/accounts/depinject.go
@@ -1,20 +1,13 @@
 package accounts
 
 import (
-	"context"
-
 	modulev1 "cosmossdk.io/api/cosmos/accounts/module/v1"
-	signingv1beta1 "cosmossdk.io/api/cosmos/tx/signing/v1beta1"
 	"cosmossdk.io/core/address"
 	"cosmossdk.io/core/appmodule"
 	"cosmossdk.io/depinject"
 	"cosmossdk.io/depinject/appconfig"
 	"cosmossdk.io/x/accounts/accountstd"
-	baseaccount "cosmossdk.io/x/accounts/defaults/base"
-	"cosmossdk.io/x/accounts/defaults/lockup"
-	"cosmossdk.io/x/accounts/defaults/multisig"
 	txdecode "cosmossdk.io/x/tx/decode"
-	"cosmossdk.io/x/tx/signing"
 
 	"github.com/cosmos/cosmos-sdk/codec"
 	cdctypes "github.com/cosmos/cosmos-sdk/codec/types"
@@ -42,8 +35,7 @@ type ModuleInputs struct {
 	Registry     cdctypes.InterfaceRegistry
 	Config       tx.ConfigOptions
 
-	// TODO: Add a way to inject custom accounts.
-	// Currently only the base account is supported.
+	Accounts []accountstd.DepinjectAccount // at least one account must be provided
 }
 
 type ModuleOutputs struct {
@@ -53,21 +45,12 @@ type ModuleOutputs struct {
 	Module         appmodule.AppModule
 }
 
-var _ signing.SignModeHandler = directHandler{}
-
-type directHandler struct{}
-
-func (s directHandler) Mode() signingv1beta1.SignMode {
-	return signingv1beta1.SignMode_SIGN_MODE_DIRECT
-}
-
-func (s directHandler) GetSignBytes(_ context.Context, _ signing.SignerData, _ signing.TxData) ([]byte, error) {
-	panic("not implemented")
-}
-
 func ProvideModule(in ModuleInputs) ModuleOutputs {
-	handler := directHandler{}
-	account := baseaccount.NewAccount("base", signing.NewHandlerMap(handler), baseaccount.WithSecp256K1PubKey())
+	accCreators := make([]accountstd.AccountCreatorFunc, len(in.Accounts))
+	for i, acc := range in.Accounts {
+		accCreators[i] = acc.MakeAccount
+	}
+
 	txDec, err := txdecode.NewDecoder(txdecode.Options{
 		SigningContext: in.Config.SigningContext,
 		ProtoCodec:     in.Cdc,
@@ -77,12 +60,8 @@ func ProvideModule(in ModuleInputs) ModuleOutputs {
 	}
 
 	accountsKeeper, err := NewKeeper(
-		in.Cdc, in.Environment, in.AddressCodec, in.Registry, txDec, account,
-		accountstd.AddAccount(lockup.CONTINUOUS_LOCKING_ACCOUNT, lockup.NewContinuousLockingAccount),
-		accountstd.AddAccount(lockup.PERIODIC_LOCKING_ACCOUNT, lockup.NewPeriodicLockingAccount),
-		accountstd.AddAccount(lockup.DELAYED_LOCKING_ACCOUNT, lockup.NewDelayedLockingAccount),
-		accountstd.AddAccount(lockup.PERMANENT_LOCKING_ACCOUNT, lockup.NewPermanentLockingAccount),
-		accountstd.AddAccount(multisig.MULTISIG_ACCOUNT, multisig.NewAccount),
+		in.Cdc, in.Environment, in.AddressCodec, in.Registry, txDec,
+		accCreators...,
 	)
 	if err != nil {
 		panic(err)