+++ x/auth/tx/config/depinject.go
@@ -49,9 +49,10 @@ type ModuleInputs struct {
 type ModuleOutputs struct {
 	depinject.Out
 
-	BaseAppOption   runtime.BaseAppOption // This is only useful for chains using baseapp.
-	TxConfig        client.TxConfig
-	TxConfigOptions tx.ConfigOptions
+	BaseAppOption       runtime.BaseAppOption // This is only useful for chains using baseapp.
+	TxConfig            client.TxConfig
+	TxConfigOptions     tx.ConfigOptions
+	TxSigningHandlerMap *txsigning.HandlerMap
 }
 
 func ProvideProtoRegistry() txsigning.ProtoFileResolver {
@@ -89,8 +90,9 @@ func ProvideModule(in ModuleInputs) ModuleOutputs {
 			app.SetTxDecoder(txConfig.TxDecoder())
 			app.SetTxEncoder(txConfig.TxEncoder())
 		},
-		TxConfig:        txConfig,
-		TxConfigOptions: txConfigOptions,
+		TxConfig:            txConfig,
+		TxConfigOptions:     txConfigOptions,
+		TxSigningHandlerMap: txConfig.SignModeHandler(),
 	}
 }